MODNOTIFY-131: maven-model 3.9.5, plexus-utils 3.5.1
Upgrade maven-model from 3.3.9 to 3.9.5, and move it to test dependency.

The upgrade indirectly upgrades plexus-utils from 3.0.22 to 3.5.1 fixing Directory Traversal and XML External Entity (XXE) Injection:

https://nvd.nist.gov/vuln/detail/CVE-2022-4244
https://nvd.nist.gov/vuln/detail/CVE-2022-4245

Because maven-model/plexus-utils is used in test code only this doesn't affect production code and no bug fix release is needed for this issue.